requirements.txt
accelerate==0.21.0
aiofiles==23.2.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.5.2
attrs==25.3.0
bitsandbytes==0.45.5
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.1.8
clip @ git+https://github.com/openai/CLIP.git@dcba3cb2e2827b402d2701e7e1c7d9fed8a20ef1
contourpy==1.1.1
cycler==0.12.1
einops==0.6.1
einops-exts==0.0.4
exceptiongroup==1.3.0
fastapi==0.115.12
ffmpy==0.5.0
filelock==3.16.1
fonttools==4.57.0
fsspec==2025.3.0
ftfy==6.2.3
gradio==4.16.0
gradio_client==0.8.1
h11==0.14.0
h5py
hf-xet==1.1.2
httpcore==0.17.3
httpx==0.24.0
huggingface-hub==0.32.0
idna==3.10
importlib_resources==6.4.5
Jinja2==3.1.6
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
markdown-it-py==3.0.0
markdown2==2.5.1
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
mpmath==1.3.0
narwhals==1.40.0
networkx==3.1
numpy 
nvidia-cublas-cu12
nvidia-cuda-cupti-cu12
nvidia-cuda-nvrtc-cu12
nvidia-cuda-runtime-cu12
nvidia-cudnn-cu12
nvidia-cufft-cu12
nvidia-curand-cu12
nvidia-cusolver-cu12
nvidia-cusparse-cu12
nvidia-nccl-cu12
nvidia-nvjitlink-cu12
nvidia-nvtx-cu12
orjson==3.10.15
packaging==25.0
pandas==2.0.3
peft==0.13.2
Pillow==9.3.0
pkgutil_resolve_name==1.3.10
protobuf==3.20.3
psutil==7.0.0
pycocotools==2.0.10
pydantic==2.10.6
pydantic_core==2.27.2
pydub==0.25.1
Pygments==2.19.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rpds-py==0.20.1
ruff==0.11.11
safetensors==0.5.3
scikit-learn==1.2.2
scipy
semantic-version==2.10.0
sentence-transformers
sentencepiece==0.1.99
shellingham==1.5.4
shortuuid==1.0.13
six==1.17.0
sniffio==1.3.1
starlette==0.44.0
svgwrite==1.4.3
sympy==1.13.3
threadpoolctl==3.5.0
timm==0.6.13
tokenizers==0.15.1
tomlkit==0.12.0
torch==2.1.2
torchvision==0.16.2
tqdm==4.67.1
transformers==4.37.2
triton
typer==0.15.4
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.2.3
uvicorn==0.33.0
wavedrom==2.0.3.post3
wcwidth==0.2.13
websockets==11.0.3
zipp==3.20.2
